refactor(knockdog): DeparturePointSheet entites ui로 이동
