--- apps/knockdog/src/features/map/index.ts
@@ -1,1 +1,0 @@
-export { useNaverOpenRoute } from './model/useNaverOpenRoute';
